fix: Don't call `get_fields_for_table` for skipped tables
Though tkutcher/quickbase-client#40 cuts down on the number of files
generated, it did not cut down on the number of times fields for tables
were requested. This commit moves the "skip" logic to before the
`get_fields_for_table` call. (It also adds pytest-mock in order to
test the change.)